Extract layout scene functions in a separate module
